// src/app/models/PartOfSpeech.ts
export enum PartOfSpeech {
  None = 'NONE',
  Verb = 'VERB', Noun = 'NOUN', Adjective = 'ADJECTIVE', Conjunction = 'CONJUNCTION',
  Postposition =  'POSTPOSITION', Adverb = 'ADVERB', Interjection = 'INTERJECTION',
  Onomatopoeia = 'ONOMATOPOEIA', Particle = 'PARTICLE', Pronoun = 'PRONOUN', Participle = 'PARTICIPLE',
  AdverbialPhrase = 'ADVERBIAL_PHRASE', AdjectivalPhrase = 'ADJECTIVAL_PHRASE', NominalPhrase = 'NOMINAL_PHRASE', VerbalPhrase = 'VERBAL_PHRASE',
  ConjunctionalPhrase = 'CONJUNCTIONAL_PHRASE',
  Phrase = 'PHRASE', Suffix = 'SUFFIX', Prefix = 'PREFIX', MultiFunctionWord = 'MULTI_FUNCTION_WORD',
  CompoundPostposition = 'COMPOUND_POSTPOSITION'
}